About the Package

SmoQyDEAC utilizes the Differential Evolution for Analytic Continuation algorithm developed by Nathan S. Nichols, Paul Sokol, and Adrian Del Maestro arXiv:2201.04155.

This package takes imaginary time correlation functions from condensed matter Monte Carlo simulations and provides the associated spectral function on the real axis.

Installation

NOTE: This package is in the experimental phase of development and is not yet published to the Julia General registry.

Open a Julia REPL environment and run the following command:

] dev https://github.com/sandimas/SmoQyDEAC.jl

This command clones the SmoQyDEAC.jl repository to the hidden direcotry .julia/dev that exists in the same directory where Julia is installed.

Running SmoQyDEAC

SmoQyDEAC has a simple API interface with a single callable function DEAC.

API

DEAC

SmoQyDEAC.DEAC — Function

DEAC(correlation_function::AbstractVector,
     correlation_function_error::AbstractVector,
     β::Float64,
     input_grid::Vector{Float64},
     out_ωs::Vector{Float64},
     kernel_type::String,
     num_bins::Int64,
     runs_per_bin::Int64,
     output_file::String,
     checkpoint_directory::String;

     population_size::Int64=8,
     base_seed::Integer=8675309,
     stop_minimum_fitness::Float64=1.0,
     number_of_generations::Int64=100000,
     keep_bin_data=true,
     autoresume_from_checkpoint=true,
     
     crossover_probability::Float64=0.9,
     self_adapting_crossover_probability::Float64=0.1,
     differential_weight::Float64=0.9,
     self_adapting_differential_weight_probability::Float64=0.1,
     self_adapting_differential_weight::Float64=0.9)

Runs the DEAC algorithm on data passed in correlation_function using $\Chi^2$ fitting from the error passed in by correlation_function_error.

Arguments

correlation_function::AbstractVector: Input data in τ space

correlation_function_error::AbstractVector: Error associated with input data

β::Float64: Inverse temperature

input_grid::Vector{Float64}: Evenly spaced values in τ from 0 to β, including end points

out_ωs::Vector{Float64}: Energies for AC output

kernel_type::String: See below for allowable kernels

num_bins::Int64: Bins to generate

runs_per_bin::Int64: Number of runs per bin for statistics

output_file::String: File to store output dictionary. jld2 format recommended

checkpoint_directory::String: Directory to store checkpoint data.

Optional Arguments

population_size::Int64=8: DEAC population size. Must be ≥ 6

base_seed::Int64=8675309: Seed

stop_minimum_fitness::Float64=1.0: Value below which fit is considered good

number_of_generations::Int64=100000: Maximum number of mutation loops

keep_bin_data::Bool=true: Save binned data or not

autoresume_from_checkpoint::Bool=true: Resume from checkpoint if possible

Optional algorithm arguments

crossover_probability::Float64=0.9: Starting likelihood of crossover

self_adapting_crossover_probability::Float64=0.1: Chance of crossover probability changing

differential_weight::Float64=0.9: Weight for second and third mutable indices

self_adapting_differential_weight_probability::Float64=0.1: Likelihood of SAD changing

self_adapting_differential_weight::Float64=0.9: SAD

Each run will use its own seed. E.g. if you run 10 bins with 100 runs per bin, you will use seeds base_seed:base_seed+999. You may increment your base seed by 1000, use another output file name, and generate more statistics later.

Output

SmoQyDEAC will save a dictionary to the file name passed via the output_file argument. The same dictionary will be returned by the function.

Checkpointing

SmoQyDEAC will place a file named DEAC_checkpoint.jld2 in the directory passed in checkpoint_directory. After completing every bin this file will be saved. After the last bin is finished the file will be deleted. When autoresume_from_checkpoint=true SmoQyDEAC will attempt to resume from the checkpoint. If the arguments passed do not match those in the checkpoint the code will exit.

Kernels

For now there are three supported kernels

time_fermionic$=\frac{1}{2}\frac{e^{-\tau\omega}}{1+e^{-\beta\omega}}$

time_bosonic$=\frac{1}{2}\frac{e^{-\tau\omega}}{1-e^{-\beta\omega}}$

time_bosonic_symmetric$=\frac{1}{2}\frac{e^{-\tau\omega}+e^{-(\beta-\tau)\omega}}{1-e^{-\beta\omega}}$

Multithreading

SmoQyDEAC utilizes Julia's Threads.@threads multithreading capability. To take advantage of this run your code using $ julia --threads=auto yourscript.jl auto will automatically use any available cores or hyperthreads. You can set the value to a fixed number as you wish.

Tips, tricks and caveats

You will likely never need to adjust any of the Optional Algorithm Arguments from the default. Many are merely initial values and will be updated stochastically early and often in the code.

The DEAC algorithm can have edge effects where it places spectral weight on the first or last ω point. This occurs when there is spectral weight just outside of your range of ωs. The solution is simply expanding the range of your output energies.

For bosonic correlations SmoQyDEAC returns the spectral function, e.g. $B(\omega)$ not $\frac{B(\omega)}{\omega}$ as some MaxEnt codes do.

Different simulation codes may report correlation functions slightly differently. E.g. for SmoQyDQMC phonon_greens $=\langle X(\tau)X(0)\rangle$ not the actual phonon green's function of $-2\Omega_0\langle X(\tau)X(0)\rangle$. While the negative sign will cancel out by our choice of Kernel, you may need to postprocess the spectral function you recover. In this case $B\rightarrow \dfrac{B}{2\Omega_0}$